Move ScalarGaussBonnet into ScalarTensor
