+++ tests/Unit/PointwiseFunctions/CMakeLists.txt
@@ -10,7 +10,6 @@ add_subdirectory(Hydro)
 add_subdirectory(InitialDataUtilities)
 add_subdirectory(MathFunctions)
 add_subdirectory(Punctures)
-add_subdirectory(ScalarGaussBonnet)
 add_subdirectory(ScalarTensor)
 add_subdirectory(SpecialRelativity)
 add_subdirectory(Xcts)
